refactor: group all protocols in a single file
